lookup connection useful for CLG activation
